Core/Achievements: phantom progress, reset, team size, death stat
- SetCriteriaProgress stored an empty progress entry before bailing out on a
  zero change, leaving null pointers behind: crash on .reset achievements and
  progress never sent to the client.
- Reset cleared the completed achievements but kept pointers to them.
- TargetInfo::Team was a uint8, so ALLIANCE (469) became 213.
- GetCriteriaTreeByType now bounds the criteria type.
- CRITERIA_TYPE_DEATH required an arena for every death statistic; only the
  arena achievements need one.

+++ src/server/game/Achievements/AchievementMgr.h
@@ -167,7 +167,7 @@ public:
         uint8 RaceID = 0;
         uint8 Level = 0;
         uint8 Gender = 0;
-        uint8 Team = 0;
+        uint32 Team = 0;
         float HealthPct = 0.0f;
         bool isAlive = true;
         bool IsMounted = false;
